Improve common testing infrastructure.
* Move from a header-only implementation to a library for
  common-test-utils.

* Implement a deferred output stream to improve parallel output
  determinism. See ctu_dprintf().
