+++ tests/common-test-utils.h
@@ -1,6 +1,6 @@
 /* -*- Mode: C++; c-basic-offset:4; indent-tabs-mode:nil -*- */
 /*
- * Copyright (c) 2020-2025 Triad National Security, LLC
+ * Copyright (c) 2020-2026 Triad National Security, LLC
  *                         All rights reserved.
  *
  * This file is part of the quo-vadis project. See the LICENSE file at the
@@ -16,6 +16,8 @@
 #ifndef COMMON_TEST_UTILS_H
 #define COMMON_TEST_UTILS_H
 
+// Avoid including internal QV headers here so
+// we mimic real application usage in our tests.
 #include "quo-vadis.h"
 
 #ifndef _GNU_SOURCE
@@ -29,42 +31,6 @@
 #include <string.h>
 #include <sys/syscall.h>
 #include <unistd.h>
-
-#ifdef __cplusplus
-
-static qvi_hwloc_bitmap
-ctu_bitmap_gen_pus(
-    size_t i,
-    size_t n
-) {
-    qvi_hwloc_bitmap result;
-    for (size_t ii = i; ii < n; ++ii) {
-        hwloc_bitmap_set(result.data(), ii);
-    }
-    return result;
-}
-
-static inline std::vector<qvi_hwloc_bitmap>
-ctu_bitmap_split_pus(
-    size_t npus,
-    size_t npieces
-) {
-    std::vector<qvi_hwloc_bitmap> result;
-    const size_t base_chunk_size = npus / npieces;
-    const size_t remainder = npus % npieces;
-    size_t current_pos = 0;
-
-    for (size_t i = 0; i < npieces; ++i) {
-        size_t current_chunk_size = base_chunk_size + (i < remainder ? 1 : 0);
-        result.emplace_back(
-            ctu_bitmap_gen_pus(current_pos, current_pos + current_chunk_size)
-        );
-        current_pos += current_chunk_size;
-    }
-    return result;
-}
-
-#endif
 
 #define CTU_STRINGIFY(x) #x
 #define CTU_TOSTRING(x)  CTU_STRINGIFY(x)
@@ -89,6 +55,10 @@ do {                                                                           \
         ctu_panic(__VA_ARGS__);                                                \
     }                                                                          \
 } while (0)
+
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 typedef enum {
     CTU_SCOPE_KIND_PROCESS = 0,
@@ -139,6 +109,17 @@ static const ctu_devid_name_to_id_t ctu_devid_name_to_id_tab[] = {
     {CTU_TOSTRING(QV_DEVICE_ID_PCI_BUS_ID), QV_DEVICE_ID_PCI_BUS_ID},
     {CTU_TOSTRING(QV_DEVICE_ID_ORDINAL),    QV_DEVICE_ID_ORDINAL}
 };
+
+// A deferred printf implementation.
+void
+ctu_dprintf(
+    const char *format,
+    ...
+);
+
+// Outputs the stored content filled by calls to ctu_dprintf() to stdout.
+void
+ctu_dflush(void);
 
 static inline const char *
 ctu_obj_name(
@@ -409,237 +390,11 @@ ctu_emit_device_info(
     printf("# -----------------------------------------------------------\n");
 }
 
-#if 0
-#define QUO_VADIS_MPI
-#include "mpi.h"
-#endif
-// We assume that the infrastructure-specific headers are included before us.
-#ifdef QUO_VADIS_MPI
-
-typedef struct {
-    MPI_Comm comm;
-    int rank;
-    int size;
-    int tok_tag;
-    bool initialized;
-} ctu_mpi_ordered_printer_t;
-
-static inline void
-ctu_mpi_ordered_printer_init(
-    ctu_scope_reporter_t *reporter,
-    qv_scope_t *scope
-) {
-    int rc = MPI_SUCCESS;
-    char *ers = NULL;
-
-    ctu_mpi_ordered_printer_t *printer = (ctu_mpi_ordered_printer_t *)reporter->printer;
-
-    do {
-        printer->initialized = false;
-        printer->tok_tag = 575;
-
-        rc = qv_mpi_scope_comm_dup(scope, &printer->comm);
-        if (rc != QV_SUCCESS) {
-            ers = "qv_mpi_scope_comm_dup";
-            rc = MPI_ERR_COMM;
-            break;
-        }
-
-        rc = MPI_Comm_rank(printer->comm, &printer->rank);
-        if (rc != MPI_SUCCESS) {
-            ers = "MPI_Comm_rank";
-            break;
-        }
-
-        rc = MPI_Comm_size(printer->comm, &printer->size);
-        if (rc != MPI_SUCCESS) {
-            ers = "MPI_Comm_size";
-            break;
-        }
-
-        printer->initialized = true;
-    } while (0);
-
-    if (rc != MPI_SUCCESS) {
-        ctu_panic("%s failed (rc=%d)", ers, rc);
-    }
-}
-
-static inline int
-ctu_mpi_ordered_printer_id(
-    ctu_scope_reporter_t *reporter
-) {
-    ctu_mpi_ordered_printer_t *printer = (ctu_mpi_ordered_printer_t *)reporter->printer;
-    return printer->rank;
-}
-
-static inline void
-ctu_mpi_ordered_printer_fini(
-    ctu_scope_reporter_t *reporter
-) {
-    int rc = MPI_SUCCESS;
-    char *ers = NULL;
-
-    ctu_mpi_ordered_printer_t *printer = (ctu_mpi_ordered_printer_t *)reporter->printer;
-
-    if (!printer || !printer->initialized) return;
-    // Ensure all pending operations are complete.
-    do {
-        rc = MPI_Barrier(printer->comm);
-        if (rc != MPI_SUCCESS) {
-            ers = "MPI_Barrier";
-            break;
-        }
-
-        rc = MPI_Comm_free(&printer->comm);
-        if (rc != MPI_SUCCESS) {
-            ers = "MPI_Comm_free";
-            break;
-        }
-        printer->initialized = false;
-        free(printer);
-    } while (0);
-    if (rc != MPI_SUCCESS) {
-        ctu_panic("%s failed (rc=%d)", ers, rc);
-    }
-}
-
-static inline void
-ctu_mpi_wait_token(
-    ctu_scope_reporter_t *reporter
-) {
-    ctu_mpi_ordered_printer_t *printer = (ctu_mpi_ordered_printer_t *)reporter->printer;
-
-    char token = 'a';
-    // Wait for token from previous rank.
-    if (printer->rank > 0) {
-        const int rc = MPI_Recv(
-            &token, 1, MPI_CHAR, printer->rank - 1,
-            printer->tok_tag, printer->comm, MPI_STATUS_IGNORE
-        );
-        if (rc != MPI_SUCCESS) {
-            ctu_panic("%s failed (rc=%d)", "MPI_Recv", rc);
-        }
-    }
-}
-
-static inline void
-ctu_mpi_release_token(
-    ctu_scope_reporter_t *reporter
-) {
-    ctu_mpi_ordered_printer_t *printer = (ctu_mpi_ordered_printer_t *)reporter->printer;
-
-    char token = 'a';
-    // Send token to next rank.
-    if (printer->rank < printer->size - 1) {
-        const int rc = MPI_Send(
-            &token, 1, MPI_CHAR, printer->rank + 1,
-            printer->tok_tag, printer->comm
-        );
-        if (rc != MPI_SUCCESS) {
-            ctu_panic("%s failed (rc=%d)", "MPI_Send", rc);
-        }
-    }
-}
-
-static inline void
-ctu_mpi_ordered_printer_vpprintf(
-    struct ctu_scope_reporter *reporter,
-    bool print,
-    const char *format,
-    va_list args
-) {
-    ctu_mpi_ordered_printer_t *printer = (ctu_mpi_ordered_printer_t *)reporter->printer;
-
-    if (!printer || !printer->initialized || !format) {
-        return;
-    }
-    // Wait for everyone to complete last operation.
-    int rc = MPI_Barrier(printer->comm);
-    if (rc != MPI_SUCCESS) {
-        ctu_panic("%s failed (rc=%d)", "MPI_Barrier", rc);
-    }
-    // Wait for my turn.
-    ctu_mpi_wait_token(reporter);
-    if (print) {
-        // Print formatted message.
-        vprintf(format, args);
-        fflush(stdout);
-    }
-    // Pass it on to the next one.
-    ctu_mpi_release_token(reporter);
-    // Wait for everyone to complete print.
-    rc = MPI_Barrier(printer->comm);
-    if (rc != MPI_SUCCESS) {
-        ctu_panic("%s failed (rc=%d)", "MPI_Barrier", rc);
-    }
-}
-
-static inline void
-ctu_mpi_ordered_printer_printf(
-    ctu_scope_reporter_t *reporter,
-    const char *format,
-    ...
-) {
-    va_list args;
-    va_start(args, format);
-    ctu_mpi_ordered_printer_vpprintf(reporter, true, format, args);
-    va_end(args);
-}
-
-static inline void
-ctu_mpi_ordered_printer_pprintf(
-    struct ctu_scope_reporter *reporter,
-    bool print,
-    const char *format,
-    ...
-) {
-    va_list args;
-    va_start(args, format);
-    ctu_mpi_ordered_printer_vpprintf(reporter, print, format, args);
-    va_end(args);
-}
-
-#endif // #ifdef QUO_VADIS_MPI
-
-static inline ctu_scope_reporter_t *
+ctu_scope_reporter_t *
 ctu_reporter_alloc(
     qv_scope_t *scope,
     ctu_scope_kind_t kind
-) {
-    ctu_scope_reporter_t *reporter = (ctu_scope_reporter_t *)calloc(1, sizeof(*reporter));
-    if (!reporter) ctu_panic("OOR!");
-
-    switch (kind) {
-        case CTU_SCOPE_KIND_PROCESS:
-        case CTU_SCOPE_KIND_PTHREAD:
-        case CTU_SCOPE_KIND_OPENMP:
-            reporter->printer = NULL;
-            reporter->init = &ctu_default_printer_init;
-            reporter->id = &ctu_default_printer_id;
-            reporter->fini = &ctu_default_printer_fini;
-            reporter->printf = &ctu_default_printer_printf;
-            reporter->pprintf = &ctu_default_printer_pprintf;
-            break;
-        case CTU_SCOPE_KIND_MPI:
-#ifdef QUO_VADIS_MPI
-            reporter->printer = calloc(1, sizeof(ctu_mpi_ordered_printer_t));
-            reporter->init = &ctu_mpi_ordered_printer_init;
-            reporter->id = &ctu_mpi_ordered_printer_id;
-            reporter->fini = &ctu_mpi_ordered_printer_fini;
-            reporter->printf = &ctu_mpi_ordered_printer_printf;
-            reporter->pprintf = &ctu_mpi_ordered_printer_pprintf;
-
-            reporter->init(reporter, scope);
-#else
-            ctu_unused(scope);
-#endif
-            break;
-        default:
-            ctu_panic("Invalid scope kind!");
-    }
-    return reporter;
-}
+);
 
 static inline void
 ctu_reporter_free(
@@ -874,6 +629,10 @@ ctu_change_bind(
 
     ctu_reporter_free(reporter);
 }
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif
