+++ tests/internal/test-map.cc
@@ -9,6 +9,38 @@
 
 #include "common-test-utils.h"
 
+static qvi_hwloc_bitmap
+ctu_bitmap_gen_pus(
+    size_t i,
+    size_t n
+) {
+    qvi_hwloc_bitmap result;
+    for (size_t ii = i; ii < n; ++ii) {
+        hwloc_bitmap_set(result.data(), ii);
+    }
+    return result;
+}
+
+static inline std::vector<qvi_hwloc_bitmap>
+ctu_bitmap_split_pus(
+    size_t npus,
+    size_t npieces
+) {
+    std::vector<qvi_hwloc_bitmap> result;
+    const size_t base_chunk_size = npus / npieces;
+    const size_t remainder = npus % npieces;
+    size_t current_pos = 0;
+
+    for (size_t i = 0; i < npieces; ++i) {
+        size_t current_chunk_size = base_chunk_size + (i < remainder ? 1 : 0);
+        result.emplace_back(
+            ctu_bitmap_gen_pus(current_pos, current_pos + current_chunk_size)
+        );
+        current_pos += current_chunk_size;
+    }
+    return result;
+}
+
 // Empty map test.
 static void
 test_1(void)
